+++ adapters/VerizonMedia/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Verizon Media iOS Mediation Adapter Changelog
 
+#### Next Version
+- Added standardized adapter error codes and messages.
+
 #### [Version 1.13.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/verizon/VerizonMediaAdapter-1.13.0.0.zip)
 - Verified compatibility with Verizon Ads Standard Edition version 1.13.0.
 - Built and tested with: